+++ src/common/components/StockSearchUnit/container.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { StockSearchUnitComponent } from './component';
+
+export const StockSearchUnit: React.FunctionComponent = () => {
+  return <StockSearchUnitComponent />;
+};