+++ src/common/components/StockSearchUnit/queries/query.ts
@@ -1,0 +1,13 @@
+import { gql } from '@apollo/client';
+
+// an example gql query
+export const GET_STOCK_BASIC = gql`
+  query GetStocksGQL($input: String) {
+    GetStocksGQL(input: $input) {
+      bestMatches {
+        name
+        symbol
+      }
+    }
+  }
+`;