+++ src/common/components/StockSearchUnit/types.ts
@@ -1,0 +1,10 @@
+export interface IBasicMatchStock {
+  name: string;
+  symbol: string;
+}
+
+export interface IBasicSearchStocks {
+  GetStocksGQL: {
+    bestMatches: IBasicMatchStock[];
+  };
+}