+++ src/pages/landing/components/LandingPage/queries/test.ts
@@ -2,10 +2,10 @@ import { gql } from '@apollo/client';
 
 // an example gql query
 export const EXCHANGE_RATES = gql`
-  query GetExchangeRates {
-    rates(currency: "USD") {
-      currency
-      rate
+  query GetBooks {
+    books {
+      title
+      author
     }
   }
 `;