+++ Changelog.md
@@ -2,5 +2,7 @@
 
 ## Unreleased
 
+- Feature: Created a simple stock search component. ([#7](https://github.com/josh-W42/stonks-viewer/pull/7))
+
 April 11th 2022
 Repo setup